Merged PR 17957: MERCH-4707: Changed SaveGlobalOptionsToDB in ROGlobalOptions
MERCH-4707: Changed SaveGlobalOptionsToDB in ROGlobalOptions to grab current Global options values from the database. This is because using _GlobalOptionsProfile for these values can create issues due to some of the defaults, for the properties, being values that you cannot properly compensate for when you are trying to save to the database.

updating UpdateGlobalOptions in Global options to make some of the parameters nullable to be more akin to the associated database parameters/columns.

Also had to update the call to UpdateGlobalOptions in windows/GlobalOptions to use the correct parameter types due to the changes.